+++ crates/store/src/lib.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 mod accounts;
 mod blocks;
 mod db;
@@ -15,6 +13,3 @@ pub use server::{DataDirectory, Store};
 // CONSTANTS
 // =================================================================================================
 const COMPONENT: &str = "miden-store";
-
-/// How often to run the database maintenance routine.
-const DATABASE_MAINTENANCE_INTERVAL: Duration = Duration::from_secs(24 * 60 * 60);